requirements.txt
Django==1.10.1
django-ckeditor==5.1.1
